fix(transport,function): close two trust-boundary truncations (review)
From the adversarial arc review (unsafe/FFI lens):

- framing write_frame / write_frame_into now reject payload >
  MAX_FRAME_SIZE_DEFAULT before the `payload.len() as u32` cast,
  mirroring the reader's existing TooLarge guard. Without it a >4 GiB
  payload would wrap the u32 length prefix and desync the stream.

- wasm.rs guest-pointer unpack uses `(packed as u64) >> 32` instead of
  the signed `(packed >> 32) as u32`, matching the documented protocol
  (upper 32 bits = pointer, always non-negative). Harmless in wasm32
  today (ptr ≤ 2^31) but removes the sign-extension footgun.

Both are pure robustness fixes, no behaviour change on valid inputs.
Gate: fmt + clippy clean; transport 32 tests, engine 1140 lib tests pass.

+++ crates/shamir-engine/src/function/wasm.rs
@@ -1231,7 +1231,7 @@ impl ShamirFunction for WasmFunction {
             .map_err(|e| map_wasm_error(e, "shamir_call"))?;
 
         // Unpack the result pointer/length.
-        let out_ptr = (packed >> 32) as u32 as usize;
+        let out_ptr = ((packed as u64) >> 32) as usize;
         let out_len = (packed & 0xFFFF_FFFF) as u32 as usize;
         let out_end = out_ptr.saturating_add(out_len);
         if out_end > memory.data_size(&store) {
